+++ lnd/src/test/java/de/cotto/bitbook/lnd/features/OnchainTransactionsServiceTest.java
@@ -1,7 +1,6 @@
 package de.cotto.bitbook.lnd.features;
 
 import de.cotto.bitbook.backend.AddressDescriptionService;
-import de.cotto.bitbook.backend.TransactionDescriptionService;
 import de.cotto.bitbook.backend.transaction.TransactionService;
 import de.cotto.bitbook.backend.transaction.model.Coins;
 import de.cotto.bitbook.backend.transaction.model.Transaction;
@@ -21,23 +20,14 @@ import static de.cotto.bitbook.backend.transaction.model.InputFixtures.INPUT_ADD
 import static de.cotto.bitbook.backend.transaction.model.InputFixtures.INPUT_ADDRESS_2;
 import static de.cotto.bitbook.backend.transaction.model.OutputFixtures.OUTPUT_ADDRESS_1;
 import static de.cotto.bitbook.backend.transaction.model.OutputFixtures.OUTPUT_ADDRESS_2;
-import static de.cotto.bitbook.backend.transaction.model.OutputFixtures.OUTPUT_VALUE_1;
 import static de.cotto.bitbook.backend.transaction.model.OutputFixtures.OUTPUT_VALUE_2;
 import static de.cotto.bitbook.backend.transaction.model.TransactionFixtures.TRANSACTION;
 import static de.cotto.bitbook.backend.transaction.model.TransactionFixtures.TRANSACTION_HASH;
 import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.FUNDING_TRANSACTION;
 import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.FUNDING_TRANSACTION_DETAILS;
+import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.ONCHAIN_TRANSACTION;
 import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.OPENING_TRANSACTION;
 import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.OPENING_TRANSACTION_DETAILS;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_CLOSE;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_CLOSE_DETAILS;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_CLOSE_EXPIRY;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_CREATION;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_CREATION_DETAILS;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_DEPOSIT;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_DEPOSIT_DETAILS;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_DEPOSIT_WITH_FEES;
-import static de.cotto.bitbook.lnd.model.OnchainTransactionFixtures.POOL_ACCOUNT_ID;
 import static de.cotto.bitbook.ownership.OwnershipStatus.OWNED;
 import static de.cotto.bitbook.ownership.OwnershipStatus.UNKNOWN;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -66,13 +56,13 @@ class OnchainTransactionsServiceTest {
     private AddressDescriptionService addressDescriptionService;
 
     @Mock
-    private TransactionDescriptionService transactionDescriptionService;
-
-    @Mock
     private TransactionService transactionService;
 
     @Mock
     private SweepTransactionsService sweepTransactionsService;
+
+    @Mock
+    private PoolTransactionService poolTransactionsService;
 
     @Test
     void loops_until_fixed_point_is_reached() {
@@ -271,212 +261,11 @@ class OnchainTransactionsServiceTest {
         }
     }
 
-    @Nested
-    class PoolAccountCreationSuccess {
-        @BeforeEach
-        void setUp() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_CREATION.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_CREATION_DETAILS);
-        }
-
-        @Test
-        void returns_number_of_accepted_transactions() {
-            assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION)))
-                    .isEqualTo(1);
-        }
-
-        @Test
-        void sets_transaction_description() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(transactionDescriptionService, atLeastOnce()).set(
-                    POOL_ACCOUNT_CREATION_DETAILS.getHash(),
-                    "Creating pool account " + POOL_ACCOUNT_ID
-            );
-        }
-
-        @Test
-        void sets_description_for_pool_address() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressDescriptionService, atLeastOnce()).set(OUTPUT_ADDRESS_2, "pool account " + POOL_ACCOUNT_ID);
-        }
-
-        @Test
-        void sets_ownership_for_pool_address() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(OUTPUT_ADDRESS_2);
-        }
-
-        @Test
-        void sets_description_for_other_outputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressDescriptionService, atLeastOnce()).set(OUTPUT_ADDRESS_1, DEFAULT_DESCRIPTION);
-        }
-
-        @Test
-        void sets_ownership_for_other_outputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(OUTPUT_ADDRESS_1);
-        }
-
-        @Test
-        void sets_ownership_for_inputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(INPUT_ADDRESS_1);
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(INPUT_ADDRESS_2);
-        }
-
-        @Test
-        void sets_descriptions_for_inputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CREATION));
-            verify(addressDescriptionService, atLeastOnce()).set(INPUT_ADDRESS_1, DEFAULT_DESCRIPTION);
-            verify(addressDescriptionService, atLeastOnce()).set(INPUT_ADDRESS_2, DEFAULT_DESCRIPTION);
-        }
-    }
-
-    @Nested
-    class PoolAccountCreationFailure {
-        @Test
-        void positive_amount() {
-            OnchainTransaction transaction = new OnchainTransaction(
-                        TRANSACTION_HASH,
-                        " poold -- AccountCreation(acct_key=" + POOL_ACCOUNT_ID + ")",
-                        Coins.ofSatoshis(123),
-                        Coins.NONE
-                );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void wrong_label() {
-            OnchainTransaction transaction = new OnchainTransaction(
-                    TRANSACTION_HASH,
-                    "poold -- AccountCreation(acct_key=" + POOL_ACCOUNT_ID + ")", // leading space missing
-                    Coins.ofSatoshis(-1_234 - 999),
-                    Coins.ofSatoshis(999)
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void mismatching_amount_for_pool_address() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_CREATION.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_CREATION_DETAILS);
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_CREATION.getTransactionHash(),
-                    POOL_ACCOUNT_CREATION.getLabel(),
-                    POOL_ACCOUNT_CREATION.getAmount().subtract(Coins.ofSatoshis(1)),
-                    POOL_ACCOUNT_CREATION.getFees()
-            );
-            assertFailure(transaction);
-        }
-    }
-
-    @Nested
-    class PoolAccountCloseSuccess {
-        @BeforeEach
-        void setUp() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_CLOSE.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_CLOSE_DETAILS);
-        }
-
-        @Test
-        void returns_number_of_accepted_transactions() {
-            assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE)))
-                    .isEqualTo(1);
-        }
-
-        @Test
-        void account_expiry() {
-            assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE_EXPIRY)))
-                    .isEqualTo(1);
-        }
-
-        @Test
-        void sets_transaction_description() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE));
-            verify(transactionDescriptionService, atLeastOnce()).set(
-                    POOL_ACCOUNT_CLOSE_DETAILS.getHash(),
-                    "Closing pool account " + POOL_ACCOUNT_ID
-            );
-        }
-
-        @Test
-        void sets_description_for_pool_addresses() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE));
-            verify(addressDescriptionService, atLeastOnce()).set(INPUT_ADDRESS_1, "pool account " + POOL_ACCOUNT_ID);
-        }
-
-        @Test
-        void sets_ownership_for_pool_addresses() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(INPUT_ADDRESS_1);
-        }
-
-        @Test
-        void sets_description_for_outputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE));
-            verify(addressDescriptionService, atLeastOnce()).set(OUTPUT_ADDRESS_1, DEFAULT_DESCRIPTION);
-        }
-
-        @Test
-        void sets_ownership_for_outputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_CLOSE));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(OUTPUT_ADDRESS_1);
-        }
-    }
-
-    @Nested
-    class PoolAccountCloseFailure {
-        @Test
-        void negative_amount() {
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_CLOSE.getTransactionHash(),
-                    POOL_ACCOUNT_CLOSE.getLabel(),
-                    Coins.ofSatoshis(-123),
-                    Coins.NONE
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void wrong_label() {
-            String label = "poold -- AccountModification(acct_key=" // leading space missing
-                           + POOL_ACCOUNT_ID
-                           + ", expiry=false, deposit=false, is_close=true)";
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_CLOSE.getTransactionHash(),
-                    label,
-                    POOL_ACCOUNT_CLOSE.getAmount(),
-                    POOL_ACCOUNT_CLOSE.getFees()
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void mismatching_amount_for_pool_address() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_CLOSE.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_CLOSE_DETAILS);
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_CLOSE.getTransactionHash(),
-                    POOL_ACCOUNT_CLOSE.getLabel(),
-                    POOL_ACCOUNT_CLOSE.getAmount().add(Coins.ofSatoshis(1)),
-                    POOL_ACCOUNT_CLOSE.getFees()
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void two_outputs() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_CLOSE.getTransactionHash()))
-                    .thenReturn(TRANSACTION);
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_CLOSE.getTransactionHash(),
-                    POOL_ACCOUNT_CLOSE.getLabel(),
-                    OUTPUT_VALUE_1.add(OUTPUT_VALUE_2),
-                    POOL_ACCOUNT_CLOSE.getFees()
-            );
-            assertFailure(transaction);
-        }
+    @Test
+    void addsForPoolTransactions() {
+        when(poolTransactionsService.addFromOnchainTransaction(any())).thenReturn(1L);
+        assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(ONCHAIN_TRANSACTION))).isEqualTo(1L);
+        verify(poolTransactionsService, atLeastOnce()).addFromOnchainTransaction(ONCHAIN_TRANSACTION);
     }
 
     @Nested
@@ -530,103 +319,6 @@ class OnchainTransactionsServiceTest {
             );
             assertFailure(transaction);
             verifyNoInteractions(sweepTransactionsService);
-        }
-    }
-
-    @Nested
-    class PoolAccountDepositSuccess {
-        @BeforeEach
-        void setUp() {
-            when(addressDescriptionService.getDescription(INPUT_ADDRESS_1)).thenReturn("");
-            when(addressDescriptionService.getDescription(INPUT_ADDRESS_2)).thenReturn(DEFAULT_DESCRIPTION);
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_DEPOSIT.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_DEPOSIT_DETAILS);
-        }
-
-        @Test
-        void returns_number_of_accepted_transactions() {
-            assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT)))
-                    .isEqualTo(1);
-        }
-
-        @Test
-        void with_fees() {
-            assertThat(onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT_WITH_FEES)))
-                    .isEqualTo(1);
-        }
-
-        @Test
-        void sets_transaction_description() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT));
-            verify(transactionDescriptionService, atLeastOnce()).set(
-                    POOL_ACCOUNT_DEPOSIT_DETAILS.getHash(),
-                    "Deposit into pool account " + POOL_ACCOUNT_ID
-            );
-        }
-
-        @Test
-        void sets_description_for_pool_addresses() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT));
-            verify(addressDescriptionService, atLeastOnce()).set(OUTPUT_ADDRESS_1, "pool account " + POOL_ACCOUNT_ID);
-        }
-
-        @Test
-        void sets_ownership_for_pool_addresses() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(OUTPUT_ADDRESS_2);
-        }
-
-        @Test
-        void sets_description_for_other_output() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT));
-            verify(addressDescriptionService, atLeastOnce()).set(OUTPUT_ADDRESS_2, DEFAULT_DESCRIPTION);
-        }
-
-        @Test
-        void sets_ownership_for_other_outputs() {
-            onchainTransactionsService.addFromOnchainTransactions(Set.of(POOL_ACCOUNT_DEPOSIT));
-            verify(addressOwnershipService, atLeastOnce()).setAddressAsOwned(OUTPUT_ADDRESS_1);
-        }
-    }
-
-    @Nested
-    class PoolAccountDepositFailure {
-        @Test
-        void nonnegative_amount() {
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_DEPOSIT.getTransactionHash(),
-                    POOL_ACCOUNT_DEPOSIT.getLabel(),
-                    Coins.ofSatoshis(123),
-                    Coins.NONE
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void wrong_label() {
-            String label = "poold -- AccountModification(acct_key=" // leading space missing
-                           + POOL_ACCOUNT_ID
-                           + ", expiry=false, deposit=true, is_close=false)";
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_DEPOSIT.getTransactionHash(),
-                    label,
-                    POOL_ACCOUNT_DEPOSIT.getAmount(),
-                    POOL_ACCOUNT_DEPOSIT.getFees()
-            );
-            assertFailure(transaction);
-        }
-
-        @Test
-        void mismatching_amount_for_pool_address() {
-            when(transactionService.getTransactionDetails(POOL_ACCOUNT_DEPOSIT.getTransactionHash()))
-                    .thenReturn(POOL_ACCOUNT_DEPOSIT_DETAILS);
-            OnchainTransaction transaction = new OnchainTransaction(
-                    POOL_ACCOUNT_DEPOSIT.getTransactionHash(),
-                    POOL_ACCOUNT_DEPOSIT.getLabel(),
-                    POOL_ACCOUNT_DEPOSIT.getAmount().add(Coins.ofSatoshis(1)),
-                    POOL_ACCOUNT_DEPOSIT.getFees()
-            );
-            assertFailure(transaction);
         }
     }
